+++ ethcore/src/client/client.rs
@@ -44,6 +44,7 @@ use receipt::LocalizedReceipt;
 pub use blockchain::CacheSize as BlockChainCacheSize;
 use trace::{TraceDB, ImportRequest as TraceImportRequest, LocalizedTrace, Database as TraceDatabase};
 use trace;
+pub use types::blockchain_info::BlockChainInfo;
 
 /// General block status
 #[derive(Debug, Eq, PartialEq)]
@@ -56,21 +57,6 @@ pub enum BlockStatus {
 	Bad,
 	/// Unknown.
 	Unknown,
-}
-
-/// Information about the blockchain gathered together.
-#[derive(Debug)]
-pub struct BlockChainInfo {
-	/// Blockchain difficulty.
-	pub total_difficulty: U256,
-	/// Block queue difficulty.
-	pub pending_total_difficulty: U256,
-	/// Genesis block hash.
-	pub genesis_hash: H256,
-	/// Best blockchain block hash.
-	pub best_block_hash: H256,
-	/// Best blockchain block number.
-	pub best_block_number: BlockNumber
 }
 
 impl fmt::Display for BlockChainInfo {
